Refactor error handler package naming.
Renamed the `error_handler` package to `error` across the codebase for consistency and simplicity. Updated all relevant imports, function calls, and references to reflect this change. No functional changes were introduced.
