+++ error/errors.go
@@ -1,4 +1,4 @@
-package error_handler
+package error
 
 import (
 	"fmt"
