+++ server/handlers.go
@@ -2,7 +2,7 @@ package server
 
 import (
 	"github.com/softwareplace/http-utils/apicontext"
-	"github.com/softwareplace/http-utils/error_handler"
+	errorhandler "github.com/softwareplace/http-utils/error"
 	"net/http"
 )
 
@@ -11,7 +11,7 @@ const ErrorHandlerWrapper = "ERROR/HANDLER/WRAPPER"
 func (a *apiRouterHandlerImpl[T]) errorHandlerWrapper(next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		ctx := apicontext.Of[T](w, r, ErrorHandlerWrapper)
-		error_handler.Handler(func() {
+		errorhandler.Handler(func() {
 			ctx.Next(next)
 		}, func(err error) {
 			if a.errorHandler != nil {
